mod_9_story/cloud_function/requirements.txt
# Function dependencies, for example:
# package>=version

functions-framework==3.*
google-cloud-logging==3.0.0
openai==0.27.2
oauth2client==4.1.3
google-cloud-ndb==2.1.1
google-auth==2.17.2
google-api-python-client==2.84.0
google-cloud-tasks==2.13.1
requests==2.28.2

